moved image modal to store container level
